Remove interface ResolverService
Because ResolverServiceImpl does not need an application context to
perform transactions, the interface ResolverService is removed and
ResolverServiceImpl has been renamed to ResolverService.
